backend/harvester/mastodon-harvest/requirements.txt
Mastodon.py
vaderSentiment
geopy
elasticsearch8
python-dateutil

